bfloat

+++ train.py
@@ -41,7 +41,7 @@ class Configuration:
     weight_decay = 0.01
 
     # model_config
-    torch_dtype = torch.float16
+    torch_dtype = torch.bfloat16
     # Only use quantization when model size is very huge because when it is
     # small and use_reentrant=False, then net gain is less then only using LORA
     load_in_8bit = False
